refactor: split memory/db.ts into db-connection, db-init, and schema-migration
Break the 2K-line monolith into focused modules:
- db-connection.ts: singleton connection (getDb/resetDb)
- db-init.ts: table creation, ALTER TABLE migrations, indexes
- schema-migration.ts: checkpoint-gated data migrations

db.ts becomes a barrel re-export so all 70+ consumers stay untouched.
